+++ volmdlr/primitives3d.py
@@ -1842,7 +1842,8 @@ class Sphere(RevolvedProfile):
 
         self.center.frame_mapping_inplace(frame, side)
 
-    def to_point_skin(self, resolution: float = 1e-3):
+    def skin_points(self, resolution: float = 1e-3):
+        """Gives points on the skin with respect to a resolution."""
         if resolution > 2 * self.radius:
             return []
 
@@ -1881,7 +1882,8 @@ class Sphere(RevolvedProfile):
 
         return skin_points
 
-    def to_point_in(self, resolution: float = 1e-3):
+    def inner_points(self, resolution: float = 1e-3):
+        """Gives points inside of the sphere with a subsphere strategy."""
         in_points = [self.center]
         nb_spheres = int(self.radius / resolution)
         if nb_spheres == 0:
@@ -1894,7 +1896,7 @@ class Sphere(RevolvedProfile):
 
         for srad in spheres_radius:
             in_sphere = Sphere(self.center, srad)
-            in_points.extend(in_sphere.to_point_skin(resolution=resolution))
+            in_points.extend(in_sphere.skin_points(resolution=resolution))
 
         return in_points